Develop

+++ src/routes/cards.routes.js
@@ -2,22 +2,24 @@ import { Router } from "express";
 import Card from "../models/Card";
 
 const router = Router();
+const { isAuthenticated } = require('../helpers/auth');
+const validateToken = require('../helpers/validateToken');
 
 //APP
 
 // Get all cards
-router.get("/cards", async (req,  res) => {
-    const cards = await Card.find().sort({createdAt: 'desc'}).lean();
+router.get("/cards", isAuthenticated, async (req,  res) => {
+    const cards = await Card.find({user: req.user.id}).sort({createdAt: 'desc'}).lean();
     res.render("cards/all-cards", { cards });
 });
 
 // Get form for adding a card
-router.get("/cards/add", (req, res) => {
+router.get("/cards/add", isAuthenticated, (req, res) => {
     res.render("cards/new-card");
 });
 
 // Post a new card
-router.post("/cards/add", async (req, res) => {
+router.post("/cards/add", isAuthenticated, async (req, res) => {
 
     var { title, description, status, priority } = req.body;
     const errors = [];
@@ -31,7 +33,7 @@ router.post("/cards/add", async (req, res) => {
         errors.push({ text: "Debe introducir un título." });
     }
     if (!status || (status != "Por hacer" && status != "En curso" && status != "Terminado")) {
-        errors.push({ text: "El estado debe ser \"Por hacer\", \"En curso\" o \"Terminado\"."});
+        errors.push({ text: "El estado debe ser 'Por hacer', 'En curso' o 'Terminado'."});
     }
     if (priority && priority != "Baja" && priority != "Media" && priority != "Alta") {
         errors.push({ text: "La prioridad debe ser baja, media o alta."});
@@ -47,6 +49,7 @@ router.post("/cards/add", async (req, res) => {
         });
     } else {
         const newCard = new Card({ title, description, status, priority });
+        newCard.user = req.user.id;
         await newCard.save();
         req.flash("success_msg", "Tarea agregada correctamente.");
         res.redirect("/cards");
@@ -54,7 +57,7 @@ router.post("/cards/add", async (req, res) => {
 });
 
 // Get form for updating a card
-router.get("/cards/edit/:id", async (req, res) => {
+router.get("/cards/edit/:id", isAuthenticated, async (req, res) => {
     const card = await Card.findById(req.params.id).lean();
     const id = req.params.id;
     const { title, description, status, priority } = card;
@@ -62,10 +65,16 @@ router.get("/cards/edit/:id", async (req, res) => {
 });
 
 // Update a card
-router.put("/cards/edit/:id", async (req, res) => {
+router.put("/cards/edit/:id", isAuthenticated, async (req, res) => {
     var { title, description, status, priority } = req.body;
     const id = req.params.id;
     const errors = [];
+
+    const card = await Card.findById(req.params.id).lean();
+
+    if(card.user !== req.user.id){
+        errors.push({ text: "No tienes permiso para editar tareas de otros usuarios."});
+    }
     if(status==""){
         status=null;
     }
@@ -76,7 +85,7 @@ router.put("/cards/edit/:id", async (req, res) => {
         errors.push({ text: "Debe introducir un título." });
     }
     if (!status || (status != "Por hacer" && status != "En curso" && status != "Terminado")) {
-        errors.push({ text: "El estado debe ser \"Por hacer\", \"En curso\" o \"Terminado\"."});
+        errors.push({ text: "El estado debe ser 'Por hacer', 'En curso' o 'Terminado'."});
     }
     if (priority && priority != "Baja" && priority != "Media" && priority != "Alta") {
         errors.push({ text: "La prioridad debe ser baja, media o alta."});
@@ -101,24 +110,29 @@ router.put("/cards/edit/:id", async (req, res) => {
 });
 
 // Delete a card
-router.delete("/cards/delete/:id", async (req, res) => {
-    await Card.findByIdAndDelete(req.params.id);
-    req.flash("success_msg", "Tarea eliminada correctamente.");
-    res.redirect("/cards");
+router.delete("/cards/delete/:id", isAuthenticated, async (req, res) => {
+    const card = await Card.findById(req.params.id).lean();
+    if(card.user !== req.user.id){
+        errors.push({ text: "No tienes permiso para eliminar tareas de otros usuarios."});
+    }else{
+        await Card.findByIdAndDelete(req.params.id);
+        req.flash("success_msg", "Tarea eliminada correctamente.");
+        res.redirect("/cards");
+    }
 });
+
 
 // API
 
 // Get all cards
-router.get("/api/cards", async (req,  res) => {
-    const cards = await Card.find().sort({createdAt: 'desc'}).lean();
+router.get("/api/cards", validateToken, async (req,  res) => {
+    const cards = await await Card.find({user: req.user.id}).sort({createdAt: 'desc'}).lean();
     res.status(200).json(cards);
 });
 
 
 // Post a new card
-
-router.post("/api/cards/add", async (req, res) => {
+router.post("/api/cards/add", validateToken, async (req, res) => {
     var { title, description, status, priority } = req.body;
 
     const errors = [];
@@ -132,7 +146,7 @@ router.post("/api/cards/add", async (req, res) => {
         errors.push({ error: "Debe introducir un título." });
     }
     if (!status || (status != "Por hacer" && status != "En curso" && status != "Terminado")) {
-        errors.push({ error: "El estado debe ser \"Por hacer\", \"En curso\" o \"Terminado\"."});
+        errors.push({ error: "El estado debe ser 'Por hacer', 'En curso' o 'Terminado'."});
     }
     if (priority && priority != "Baja" && priority != "Media" && priority != "Alta") {
         errors.push({ error: "La prioridad debe ser baja, media o alta."});
@@ -142,6 +156,7 @@ router.post("/api/cards/add", async (req, res) => {
         res.status(400).json(errors);
     } else {
         const newCard = new Card({ title, description, status, priority });
+        newCard.user = req.user.id;
         await newCard.save();
         res.status(201).json(newCard);
     }
@@ -149,11 +164,16 @@ router.post("/api/cards/add", async (req, res) => {
 });
 
 // Update a card
-
-router.put("/api/cards/edit/:id", async (req, res) => {
+router.put("/api/cards/edit/:id", validateToken, async (req, res) => {
     var { title, description, status, priority } = req.body;
     const id = req.params.id;
     const errors = [];
+
+    const card = await Card.findById(req.params.id).lean();
+
+    if(card.user !== req.user.id){
+        errors.push({ error: "No tienes permiso para editar tareas de otros usuarios."});
+    }
     if(status==""){
         status=null;
     }
@@ -164,7 +184,7 @@ router.put("/api/cards/edit/:id", async (req, res) => {
         errors.push({ error: "Debe introducir un título." });
     }
     if (!status || (status != "Por hacer" && status != "En curso" && status != "Terminado")) {
-        errors.push({ error: "El estado debe ser \"Por hacer\", \"En curso\" o \"Terminado\"."});
+        errors.push({ error: "El estado debe ser 'Por hacer', 'En curso' o 'Terminado'."});
     }
     if (priority && priority != "Baja" && priority != "Media" && priority != "Alta") {
         errors.push({ error: "La prioridad debe ser baja, media o alta."});
@@ -179,9 +199,14 @@ router.put("/api/cards/edit/:id", async (req, res) => {
 });
 
 // Delete a card
-router.delete("/api/cards/delete/:id", async (req, res) => {
-    await Card.findByIdAndDelete(req.params.id);
-    res.status(200).json({ message:"Tarea eliminada correctamente" })
+router.delete("/api/cards/delete/:id", validateToken, async (req, res) => {
+    const card = await Card.findById(req.params.id).lean();
+    if(card.user !== req.user.id){
+        res.status(403).json({ error:"No tienes permisos para eliminar tareas de otro usuario" });
+    }else{
+        await Card.findByIdAndDelete(req.params.id);
+        res.status(200).json({ message:"Tarea eliminada correctamente" })
+    }
 });
 
 export default router;
